:sparkles: Added BranchController and fix status to default in migration

+++ app/Http/Controllers/API/BranchController.php
@@ -2,55 +2,92 @@
 
 namespace App\Http\Controllers\API;
 
-use App\Models\Branch;
-use Illuminate\Http\Request;
-use App\Http\Requests\BranchRequest;
-use Illuminate\Http\Response;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\BranchRequest;
 use App\Http\Resources\BranchResource;
+use App\Models\Branch;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
+use Illuminate\Http\Response;
+use Symfony\Component\HttpFoundation\Response as HttpResponse;
 
 class BranchController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+
     public function index(Request $request)
     {
-        $branches = Branch::paginate();
+        if ($request->brand_id != null) {
+            $branches = Branch::where('brand_id', $request->brand_id)->paginate();
 
-        return BranchResource::collection($branches);
+            return BranchResource::collection($branches);
+        }
+
+        return response()->json(['message' => 'Brand ID is required.'], HttpResponse::HTTP_BAD_REQUEST);
+
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(BranchRequest $request): Branch
+    public function privateIndex(Request $request)
     {
-        return Branch::create($request->validated());
+        if (!$this->validateUser()) {
+            return response()->json(['error' => 'Unauthorized'], HttpResponse::HTTP_UNAUTHORIZED);
+        }
+
+        $branches = Branch::where('user_id', auth()->id())->get();
+
+        return response()->json($branches);
     }
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(Branch $branch): Branch
+    public function store(BranchRequest $request): JsonResponse
     {
-        return $branch;
+        if (!$this->validateUser()) {
+            return response()->json(['error' => 'Unauthorized'], HttpResponse::HTTP_BAD_REQUEST);
+        }
+
+        return response()->json([Branch::create($request->validated()), HttpResponse::HTTP_CREATED]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(BranchRequest $request, Branch $branch): Branch
+    public function show(int $id): JsonResponse
+    {
+        $branch = Branch::find($id);
+
+        if ($branch == null) {
+            return response()->json(['message' => 'Branch not found.'], HttpResponse::HTTP_NOT_FOUND);
+        }
+
+        return response()->json($branch);
+    }
+
+
+    public function update(BranchRequest $request, Branch $branch): JsonResponse
     {
         $branch->update($request->validated());
 
-        return $branch;
+        return response()->json($branch);
     }
 
-    public function destroy(Branch $branch): Response
+    public function destroy(int $id): JsonResponse
     {
+        if (!$this->validateUser()) {
+            return response()->json(['error' => 'Unauthorized'], HttpResponse::HTTP_UNAUTHORIZED);
+        }
+
+        $branch = Branch::find($id);
+        if (!isset($branch)) {
+            return response()->json(['error' => 'Not found'], HttpResponse::HTTP_NOT_FOUND);
+        }
+
         $branch->delete();
 
         return response()->noContent();
+    }
+
+    protected function validateUser(): bool
+    {
+        // Check if the authenticated user has the necessary role
+        if (auth()->user()->type == "admin" || auth()->user()->type == "Superadmin") {
+            return true;
+        }
+        return false;
+
     }
 }
